+++ docs/changelog.md
@@ -2,7 +2,7 @@
 
 This project attempts to follow [Semantic Versioning](https://semver.org) and uses [Keep-a-Changelog formatting](https://keepachangelog.com/en/1.0.0/). But I make mistakes sometimes.
 
-## [3.8.0]
+## [4.8.0]
 
 ### Added